# Optimize JsonWriter APIs be separating prettyPrint code paths