``` ini

BenchmarkDotNet=v0.10.14.534-nightly, OS=Windows 10.0.14393.2312 (1607/AnniversaryUpdate/Redstone1)
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
Frequency=3328126 Hz, Resolution=300.4694 ns, Timer=TSC
.NET Core SDK=2.2.100-preview1-009015
  [Host]     : .NET Core 2.2.0-preview1-26609-02 (CoreCLR 4.6.26606.04, CoreFX 4.6.26606.04), 64bit RyuJIT
  DefaultJob : .NET Core 2.2.0-preview1-26609-02 (CoreCLR 4.6.26606.04, CoreFX 4.6.26606.04), 64bit RyuJIT


```

**Before (master):**

|                         Method | Formatted | size |         Mean |      Error |     StdDev |       Median | Allocated |
|------------------------------- |---------- |----- |-------------:|-----------:|-----------:|-------------:|----------:|
|  **WriterSystemTextJsonArrayOnly** |     **False** |    **1** |     **79.62 ns** |   **1.593 ns** |   **2.336 ns** |     **79.02 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** |   **10** |    **170.43 ns** |   **1.966 ns** |   **1.642 ns** |    **169.84 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** | **1000** | **13,846.95 ns** | **276.462 ns** | **761.456 ns** | **13,584.43 ns** |       **0 B** |
|      **WriterSystemTextJsonBasic** |     **False** |    **?** |    **825.65 ns** |  **13.743 ns** |  **12.855 ns** |    **823.55 ns** |       **0 B** |
| WriterSystemTextJsonHelloWorld |     False |    ? |    105.45 ns |   2.087 ns |   2.856 ns |    105.69 ns |       0 B |
|  **WriterSystemTextJsonArrayOnly** |      **True** |    **1** |     **89.78 ns** |   **1.813 ns** |   **3.223 ns** |     **89.03 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |      **True** |   **10** |    **210.94 ns** |   **4.393 ns** |   **7.218 ns** |    **208.56 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |      **True** | **1000** | **16,035.17 ns** | **336.538 ns** | **679.823 ns** | **15,869.11 ns** |       **0 B** |
|      **WriterSystemTextJsonBasic** |      **True** |    **?** |    **934.21 ns** |  **16.576 ns** |  **17.022 ns** |    **930.61 ns** |       **0 B** |
| WriterSystemTextJsonHelloWorld |      True |    ? |    116.00 ns |   1.707 ns |   1.513 ns |    116.06 ns |       0 B |

**After (this PR):**

|                         Method | Formatted | size |         Mean |       Error |      StdDev | Allocated |
|------------------------------- |---------- |----- |-------------:|------------:|------------:|----------:|
|  **WriterSystemTextJsonArrayOnly** |     **False** |    **1** |     **76.77 ns** |   **1.1029 ns** |   **1.0317 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** |   **10** |    **166.06 ns** |   **1.5836 ns** |   **1.4038 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** | **1000** | **13,767.79 ns** | **275.3813 ns** | **530.5663 ns** |       **0 B** |
|      **WriterSystemTextJsonBasic** |     **False** |    **?** |    **787.15 ns** |  **15.6772 ns** |  **27.0425 ns** |       **0 B** |
| WriterSystemTextJsonHelloWorld |     False |    ? |     96.45 ns |   1.9487 ns |   2.0851 ns |       0 B |
|  **WriterSystemTextJsonArrayOnly** |      **True** |    **1** |     **91.84 ns** |   **1.8149 ns** |   **1.8638 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |      **True** |   **10** |    **214.31 ns** |   **2.6148 ns** |   **2.0414 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |      **True** | **1000** | **16,959.28 ns** |  **93.3465 ns** |  **72.8788 ns** |       **0 B** |
|      **WriterSystemTextJsonBasic** |      **True** |    **?** |    **907.49 ns** |  **17.3642 ns** |  **17.0539 ns** |       **0 B** |
| WriterSystemTextJsonHelloWorld |      True |    ? |    108.39 ns |   0.3011 ns |   0.2816 ns |       0 B |


cc @benaadams, @GrabYourPitchforks