@benaadams 2018-06-28
Nice :)

Interesting that both `Basic` and `HelloWorld` improve for formatted; would guess the due to the inlining parent method, keeping it at the same call depth. Regression in other formatted seems valid as if you want the fastest performance, you probably aren't going to want to add lots of whitespace.