@ahsonkhan 2018-06-29
> Interesting that both `Basic` and `HelloWorld` improve for formatted; would guess the due to the inlining parent method, keeping it at the same call depth. 

Generally speaking, I think that is within the margin of error (I observed small variations in the numbers between runs depending on what I am doing on my machine).

cc @adamsitnik