@adamsitnik 2018-06-29
> Generally speaking, I think that is within the margin of error (I observed small variations in the numbers between runs depending on what I am doing on my machine).

It's recommended to close at least the most CPU consuming apps before running the "final" benchmarks. I personally just close VS/Rider and Chrome.
The other thing that can help is to set the processor affinity `job.WithAffinity(mask)`